requirements.txt
openai-whisper
torch
gradio
google-generativeai